+++ Changelog.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented here.
 
 ## [unreleased]
 - Update Python and Jupyter test names to follow `[file] class.funcname` format (#605)
+- Update r tester test_name formatting and add r tester tests (#606)
 
 ## [v2.7.0]
 - Update python, pyta and jupyter testers to allow a requirements file (#580)