+++ Changelog.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented here.
 
 ## [unreleased]
-- Add `ai_tester` module to support AI-based autograding via ai-autograding-feedback (#620)
+- Update Python and Jupyter test names to follow `[file] class.funcname` format (#605)
+- Update r tester test_name formatting and add r tester tests (#606)
+- Update R tester to support MarkUs metadata (#615)
+- Update Python tester to support MarkUs metadata for test-specific marks earned and marks total (#619)
+- Add `ai_tester` module to support AI-based autograding via `ai-autograding-feedback` (#620)
+
 
 ## [v2.7.0]
 - Update python, pyta and jupyter testers to allow a requirements file (#580)