+++ Changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented here.
 - Update Python tester to support MarkUs metadata for test-specific marks earned and marks total (#619)
 - Add `ai_tester` module to support AI-based autograding via `ai-autograding-feedback` (#625)
 - Add `ai` to list of testers (#628)
+- Fixed an `AttributeError` when handling exceptions in server `update_test_settings` (#629)
 
 ## [v2.7.0]
 - Update python, pyta and jupyter testers to allow a requirements file (#580)