+++ Changelog.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented here.
 - Add `ai_tester` module to support AI-based autograding via `ai-autograding-feedback` (#625)
 - Add `ai` to list of testers (#628)
 - Fixed an `AttributeError` when handling exceptions in server `update_test_settings` (#629)
+- Modified R tester to always display test result messages, even when tests pass (#633)
 
 ## [v2.7.0]
 - Update python, pyta and jupyter testers to allow a requirements file (#580)