+++ Changelog.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented here.
 - Add `ai` to list of testers (#628)
 - Fixed an `AttributeError` when handling exceptions in server `update_test_settings` (#629)
 - Add tag functionalty to `ai_tester` (#631)
+- Modified R tester to always display test result messages, even when tests pass (#633)
 
 ## [v2.7.0]
 - Update python, pyta and jupyter testers to allow a requirements file (#580)